+++ src/widgets/ResultItemWidget.cpp
@@ -1,5 +1,6 @@
 #include "ResultItemWidget.h"
 #include "../core/HistoryManager.h"
+#include "../common/ResultItem.h"
 
 ResultItemWidget::ResultItemWidget(QListWidget* parent) : QListWidgetItem(parent) {}
 
@@ -7,8 +8,8 @@ bool ResultItemWidget::operator<(const QListWidgetItem& other) const
 {
     const int thisPriority = this->data(Qt::UserRole).value<ResultItem>().priority;
     const int otherPriority = other.data(Qt::UserRole).value<ResultItem>().priority;
-    const int thisScore = this->data(Qt::UserRole).value<ResultItem>().score;
-    const int otherScore = other.data(Qt::UserRole).value<ResultItem>().score;
+    const double thisScore = this->data(Qt::UserRole).value<ResultItem>().score;
+    const double otherScore = other.data(Qt::UserRole).value<ResultItem>().score;
     const QString thisKey = this->data(Qt::UserRole).value<ResultItem>().key;
     const QString otherKey = other.data(Qt::UserRole).value<ResultItem>().key;
     const double thisHistoryScore = HistoryManager::getHistoryScore(thisKey);
